--- packages/ui-components/src/components/Maps/RenderMapProps.ts
@@ -1,6 +1,0 @@
-export interface RenderMapProps {
-  renderTooltip: (regionId: string) => React.ReactElement | string;
-  getFillColor?: (regionId: string) => string;
-  getRegionUrl?: (regionId: string) => string | undefined;
-  width?: number;
-}
