--- packages/ui-components/src/components/Maps/USNationalMap/index.ts
@@ -1,3 +1,0 @@
-export * from "./USNationalMap"; // USNationalMap
-export * from "./MetricUSNationalMap"; // MetricUSNationalMap
-export * from "./interfaces"; // USNationalMapProps, MetricUSNationalMapProps
