--- packages/ui-components/src/components/Maps/USStateMap/index.ts
@@ -1,3 +1,0 @@
-export * from "./USStateMap"; // USStateMap
-export * from "./MetricUSStateMap"; // MetricUSStateMap
-export * from "./interfaces"; // USStateMapProps, MetricUSStateMapProps
