+++ packages/ui-components/src/components/MetricUSMaps/MetricUSNationalMap/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetricUSNationalMap";
