+++ packages/ui-components/src/components/MetricUSMaps/MetricUSStateMap/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetricUSStateMap";
