+++ packages/ui-components/src/components/MetricUSMaps/index.ts
@@ -1,0 +1,2 @@
+export * from "./MetricUSNationalMap";
+export * from "./MetricUSStateMap";
