+++ packages/ui-components/src/components/USMaps/USNationalMap/index.ts
@@ -1,0 +1,2 @@
+export * from "./USNationalMap";
+export * from "../interfaces";
