+++ packages/ui-components/src/components/USMaps/USStateMap/index.ts
@@ -1,0 +1,1 @@
+export * from "./USStateMap";
