+++ packages/ui-components/src/components/USMaps/index.ts
@@ -1,2 +1,3 @@
 export * from "./USNationalMap";
 export * from "./USStateMap";
+export * from "./interfaces";
