+++ packages/ui-components/src/index.ts
@@ -89,7 +89,7 @@ export * from "./components/LabelIcon";
 export * from "./components/LegendCategorical";
 export * from "./components/LegendThreshold";
 export * from "./components/LineChart";
-export * from "./components/Maps";
+export * from "./components/USMaps";
 export * from "./components/Markdown";
 export * from "./components/MetricCatalogContext";
 export * from "./components/MetricCompareTable";
